Extract classes into separate files
