+++ Source/Core/AST/Commands/HideRevealCmd.cs
