+++ Source/Core/AST/GotoBoogie/ChangeScope.cs
