+++ Source/Core/AST/StructuredBoogie/StmtListBuilder.cs
@@ -1,0 +1,98 @@
+using System.Collections.Generic;
+using System.Diagnostics.Contracts;
+
+namespace Microsoft.Boogie;
+
+/// <summary>
+/// The AST for Boogie structured commands was designed to support backward compatibility with
+/// the Boogie unstructured commands.  This has made the structured commands hard to construct.
+/// The StmtListBuilder class makes it easier to build structured commands.
+/// </summary>
+public class StmtListBuilder
+{
+  readonly List<BigBlock /*!*/> /*!*/ bigBlocks = new();
+
+  string label;
+  List<Cmd> simpleCmds;
+
+  [ContractInvariantMethod]
+  void ObjectInvariant()
+  {
+    Contract.Invariant(cce.NonNullElements(bigBlocks));
+  }
+
+  void Dump(StructuredCmd scmd, TransferCmd tcmd)
+  {
+    Contract.Requires(scmd == null || tcmd == null);
+    Contract.Ensures(label == null && simpleCmds == null);
+    if (label == null && simpleCmds == null && scmd == null && tcmd == null)
+    {
+      // nothing to do
+    }
+    else
+    {
+      if (simpleCmds == null)
+      {
+        simpleCmds = new List<Cmd>();
+      }
+
+      bigBlocks.Add(new BigBlock(Token.NoToken, label, simpleCmds, scmd, tcmd));
+      label = null;
+      simpleCmds = null;
+    }
+  }
+
+  /// <summary>
+  /// Collects the StmtList built so far and returns it.  The StmtListBuilder should no longer
+  /// be used once this method has been invoked.
+  /// </summary>
+  public StmtList Collect(IToken endCurlyBrace)
+  {
+    Contract.Requires(endCurlyBrace != null);
+    Contract.Ensures(Contract.Result<StmtList>() != null);
+    Dump(null, null);
+    if (bigBlocks.Count == 0)
+    {
+      simpleCmds = new List<Cmd>(); // the StmtList constructor doesn't like an empty list of BigBlock's
+      Dump(null, null);
+    }
+
+    return new StmtList(bigBlocks, endCurlyBrace);
+  }
+
+  public void Add(Cmd cmd)
+  {
+    Contract.Requires(cmd != null);
+    if (simpleCmds == null)
+    {
+      simpleCmds = new List<Cmd>();
+    }
+
+    simpleCmds.Add(cmd);
+  }
+
+  public void Add(StructuredCmd scmd)
+  {
+    Contract.Requires(scmd != null);
+    Dump(scmd, null);
+  }
+
+  public void Add(TransferCmd tcmd)
+  {
+    Contract.Requires(tcmd != null);
+    Dump(null, tcmd);
+  }
+
+  public void AddLabelCmd(string label)
+  {
+    Contract.Requires(label != null);
+    Dump(null, null);
+    this.label = label;
+  }
+
+  public void AddLocalVariable(string name)
+  {
+    Contract.Requires(name != null);
+    // TODO
+  }
+}
